pyproject.toml
[project]
name = "hugmuninn"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "discord-py>=2.7.0",
    "fastapi>=0.135.1",
    "google-api-python-client>=2.190.0",
    "google-auth-httplib2>=0.3.0",
    "google-auth-oauthlib>=1.3.0",
    "langchain-core>=1.2.16",
    "langchain-ollama>=1.0.1",
    "langchain-tavily>=0.2.17",
    "langgraph>=1.0.10",
    "playwright>=1.58.0",
    "python-dotenv>=1.2.2",
]
